git-svn-id: svn://123.57.173.36/jhj-parent2.0/Branches/jhj-parent2.1@405 57d5540f-7636-4a73-8af3-7dc9f519f443

+++ jhj-dao/src/main/java/com/jhj/po/dao/user/UserAddrsMapper.java
@@ -12,7 +12,7 @@ public interface UserAddrsMapper {
 	
 	    Long insert(UserAddrs record);
 	
-	    int insertSelective(UserAddrs record);
+	    Long insertSelective(UserAddrs record);
 	
 	    int updateByPrimaryKeySelective(UserAddrs record);
